pyproject.toml
[tool.poetry]
name = "foodgram-project-react"
version = "0.1.0"
description = ""
authors = ["Каманин Юрий <kamanchi22@mail.ru>"]
readme = "README.md"
packages = [{include = "foodgram_project_react"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14.0"
black = "^23.3.0"
poetry = "^1.4.2"
pillow = "^9.5.0"
django-filter = "^23.2"
django-shortcuts = "^1.6"
drf-writable-nested = "^0.7.0"
psycopg2 = "^2.9.6"
drf-extra-fields = "^3.4.1"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
